pyproject.toml
[project]
name = "kellasor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.8.1",
    "async-timeout==5.0.1",
    "celery>=5.6.2",
    "django==4.2.20",
    "django-filter==25.1",
    "django-redis==5.4.0",
    "djangorestframework==3.16.0",
    "djangorestframework-simplejwt==5.3.1",
    "faker==37.1.0",
    "psycopg>=3.3.2",
    "pyjwt==2.10.1",
    "python-dotenv==1.1.0",
    "redis==6.0.0",
    "regex==2024.11.6",
    "ruff>=0.14.14",
    "sqlparse==0.5.3",
    "typing-extensions==4.13.2",
    "tzdata==2025.2",
]
